+++ SDK-sources/OutbrainSDK/SmartFeed/SmartFeedManager/SmartFeedManager.m
@@ -445,7 +445,6 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
     // build the index paths for insertion
     // since you're adding to the end of datasource, the new rows will start at count + 1 (header)
     NSInteger baseIndex = self.smartFeedItemsArray.count+1;
-    
     for (int i = 0; i < newSmartfeedItems.count; i++) {
         [indexPaths addObject:[NSIndexPath indexPathForRow:baseIndex+i inSection:self.outbrainSectionIndex]];
     }
@@ -479,6 +478,14 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
         if (self.sfTableViewManager.tableView != nil) {
             // tell the table view to update (at all of the inserted index paths)
             UITableView *tableView = self.sfTableViewManager.tableView;
+            
+            // Check if Sky solution is needed
+            BOOL isSkySolutionActive = [self isSkySolutionActive:tableView baseIndex:baseIndex];
+            if (isSkySolutionActive) {
+                [self skySolutionForTableViewReload:tableView newSmartfeedItems:newSmartfeedItems indexPaths:indexPaths];
+                return;
+            }
+            
             [tableView beginUpdates];
             [self.smartFeedItemsArray addObjectsFromArray:newSmartfeedItems];
             [tableView insertRowsAtIndexPaths:indexPaths withRowAnimation:UITableViewRowAnimationNone];
@@ -491,6 +498,10 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
     if (indexPath.section != self.outbrainSectionIndex) {
         return nil;
+    }
+    
+    if (indexPath.row >= [self smartFeedItemsCount]) {
+        return [[UITableViewCell alloc] init];
     }
     
     if (indexPath.row == 0) {
@@ -541,6 +552,10 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
         return;
     }
     
+    if (indexPath.row >= [self smartFeedItemsCount]) {
+        return;
+    }
+    
     if (indexPath.row == 0) {
         // Smartfeed header
         [self configureSmartFeedHeaderTableViewCell:cell atIndexPath:indexPath];
@@ -582,6 +597,10 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
         // Smartfeed header
         return UITableViewAutomaticDimension;
     }
+    if (indexPath.row >= [self smartFeedItemsCount]) {
+        return 0.1;
+    }
+    
     SFItemData *sfItem = [self itemForIndexPath:indexPath];
     return [self.sfTableViewManager heightForRowAtIndexPath:indexPath withSFItem:sfItem];
 }
@@ -726,6 +745,28 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
     UITapGestureRecognizer *tapGesture = [[UITapGestureRecognizer alloc] initWithTarget:self  action:@selector(outbrainLabelClicked:)];
     tapGesture.numberOfTapsRequired = 1;
     [sfHeaderCell.contentView addGestureRecognizer:tapGesture];
+}
+
+- (BOOL) isSkySolutionActive:(UITableView *)tableView baseIndex:(NSInteger)baseIndex {
+    NSInteger currentNumberOfSections = [tableView numberOfSections];
+    NSInteger currentNumberOfItemsInSection = [tableView numberOfRowsInSection:self.outbrainSectionIndex];
+    return self.outbrainSectionIndex < currentNumberOfSections && baseIndex < currentNumberOfItemsInSection;
+}
+
+- (void) skySolutionForTableViewReload:(UITableView *)tableView newSmartfeedItems:(NSArray *)newSmartfeedItems indexPaths:(NSArray *)indexPaths {
+    [self.smartFeedItemsArray addObjectsFromArray:newSmartfeedItems];
+    
+    // Check if there is an overlap between visibleIndexPathArray and the new IndexPaths we are about to add.
+    NSArray *visibleIndexPathArray = [tableView indexPathsForVisibleRows];
+    NSMutableSet *set1 = [NSMutableSet setWithArray: visibleIndexPathArray];
+    NSSet *set2 = [NSSet setWithArray: indexPaths];
+    [set1 intersectSet: set2];
+    NSArray *resultArray = [set1 allObjects];
+    
+    if ([resultArray count] > 1) {
+        // Detected overlap - calling reloadRowsAtIndexPaths (this will initiate a call to heightForRowAt:
+        [tableView reloadRowsAtIndexPaths:indexPaths withRowAnimation:UITableViewRowAnimationAutomatic];
+    }
 }
 
 #pragma mark - Collection View methods
@@ -971,6 +1012,10 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
         return SFTypeSmartfeedHeader;
     }
     
+    if (indexPath.row >= [self smartFeedItemsCount]) {
+        return SFTypeBadType;
+    }
+    
     SFItemData *sfItem = [self itemForIndexPath:indexPath];
     return sfItem.itemType;
 }
@@ -1019,6 +1064,10 @@ int const OBVIEW_DEFAULT_TAG = 12345678;
     
     if (indexPath.row == 0) {
         // Smartfeed header        
+        return nil;
+    }
+    
+    if (indexPath.row >= [self smartFeedItemsCount]) {
         return nil;
     }